requirements.txt
PyPDF2==3.0.1
reportlab==4.0.7
PyYAML~=6.0.1
translate~=3.6.1